use new gnutls version 3.6.11.1

+++ src/gnutls/src/libopts/time.c
@@ -8,7 +8,7 @@
 /*
  *  This file is part of AutoOpts, a companion to AutoGen.
  *  AutoOpts is free software.
- *  AutoOpts is Copyright (C) 1992-2015 by Bruce Korb - all rights reserved
+ *  AutoOpts is Copyright (C) 1992-2018 by Bruce Korb - all rights reserved
  *
  *  AutoOpts is available under any one of two licenses.  The license
  *  in use must be one of these two and the choice is under the control
@@ -92,7 +92,7 @@ optionTimeDate(tOptions * opts, tOptDesc * od)
             static char const fmt[] = "DATEMSK=%s/datemsk";
             size_t sz = sizeof(fmt) + strlen(opts->pzPkgDataDir);
             envptr = AGALOC(sz, fmt);
-            if (snprintf(envptr, sz, fmt, opts->pzPkgDataDir) >= sz)
+            if (snprintf(envptr, sz, fmt, opts->pzPkgDataDir) >= (int)sz)
                 option_exits(EXIT_FAILURE);
 
             putenv(envptr);
